refactor: move roles to constants.ts file
