+++ server/src/common/constants.ts
@@ -89,3 +89,10 @@ export const TRAINING_RYTHM = {
   "2S3S": "2 semaines / 3 semaines",
   "6S6S": "6 semaines / 6 semaines",
 }
+
+export const roles = {
+  candidat: "candidat",
+  cfa: "cfa",
+  administrator: "administrator",
+}
+
