--- server/src/common/roles.ts
@@ -1,5 +1,0 @@
-export const roles = {
-  candidat: "candidat",
-  cfa: "cfa",
-  administrator: "administrator",
-}
