+++ server/src/http/routes/appointmentRequest.controller.ts
@@ -5,7 +5,7 @@ import { mailTemplate } from "../../assets/index.js"
 import * as appointmentService from "../../services/appointment.service.js"
 import { mailType } from "../../common/model/constants/appointments.js"
 import { getReferrerByKeyName } from "../../common/model/constants/referrers.js"
-import { roles } from "../../common/roles.js"
+import { roles } from "../../common/constants.js"
 import { dayjs } from "../../common/utils/dayjs.js"
 import config from "../../config.js"
 import { tryCatch } from "../middlewares/tryCatchMiddleware.js"
