+++ server/src/jobs/rdv/oneTimeJob/addLastActionDateToUserCollection.ts
@@ -1,6 +1,6 @@
 import dayjs from "../../../common/dayjs.js"
 import { User } from "../../../common/model/index.js"
-import { roles } from "../../../common/roles.js"
+import { roles } from "../../../common/constants.js"
 import { runScript } from "../../scriptWrapper.js"
 
 export const addLastActionDateToUserCollection = async () => {
