+++ server/tests/integration/http/passwordRoutes-test.ts
@@ -4,7 +4,7 @@ import jwt from "jsonwebtoken"
 import config from "../../../src/config.js"
 import httpTests from "../../utils/httpTests.js"
 import { createPasswordToken } from "../../../src/common/utils/jwtUtils.js"
-import { roles } from "../../../src/common/roles.js"
+import { roles } from "../../../src/common/constants.js"
 import __filename from "../../../src/common/filename.js"
 
 httpTests(__filename(import.meta.url), ({ startServer }) => {
